==> scripts/1_tidy_production_data.Rmd <==
---
title: "Tidying FAO production data"
output: html_document
---

##Libraries and setup

```{r setup, include=FALSE}

library(janitor) 
library(here)
library(tidyverse)
library(vroom)
library(countrycode)

source(here("source", "directories.R"))

```


##Import livestock production data (stocks and primary)

```{r, include=FALSE}

#Stocks
livestock_animals <- vroom(here(dir_faostat_2021,  "Production_Livestock_E_All_Data_(Normalized).csv")) %>% 
  clean_names()

#bit o' exploration
unique(livestock_animals$item)
unique(livestock_animals$element) #only stocks
unique(livestock_animals$unit)

livestock_animals %>% filter(unit=="No") %>% .$item %>% unique() #items for unit in number "No." is beehives 
(head_1000 <- livestock_animals %>% filter(unit=="1000 Head") %>% .$item %>% unique()) #1000 Head mainly for Poultry and rodents
(head <- livestock_animals %>% filter(unit=="Head") %>% .$item %>% unique()) #1000 Head mainly for Poultry and rodents


#Primary commodities

livestock_primary <- vroom(here(dir_faostat_2021,  "Production_LivestockPrimary_E_All_Data_(Normalized).csv")) %>% 
  clean_names() 

#bit o' exploration
sort(unique(livestock_primary$item))
sort(unique(livestock_primary$element))
unique(livestock_primary$unit)

livestock_primary %>% filter(element == "Prod Popultn") %>% .$item %>% unique()  #production population is for wool, honey,beeswax
livestock_primary %>% filter(element=="Producing Animals/Slaughtered") %>% .$item %>% unique() #for all commodities but the stocks are probably more valuable here - several reasons producers might not slaughter but if stocks are increasing this can hide trends
livestock_primary %>% filter(element=="Production") %>% .$item %>% unique() # all products have production data
livestock_primary %>% filter(element == "Production") %>% .$unit %>% unique() # we want tonnes but what about eggs
livestock_primary %>% filter(element == "Production" & unit== "1000 No") %>% .$item %>% unique() # eggs in numbers
livestock_primary %>% filter(element == "Production" & unit== "tonnes") %>% .$item %>% unique() # eggs in numbers and in tonnes


```



##Select and group items and elements we want

```{r, include=FALSE}

#Stocks data (mainly about choosing items)

(
  stocks_tidy <- livestock_animals %>% 
    #take out elements we don't want  
    filter(item %in% c("Cattle", "Buffaloes", "Chickens", "Sheep", "Goats", "Pigs" )) %>% 
    #aggregate some groups e.g. guinea fowl
    mutate(group = case_when(item %in% c("Buffaloes", "Sheep",  "Goats") ~ "Other",
                             TRUE ~ item)) %>% 
    #adjust poultry and rodents value to 1000 fold greater because of reported units
    mutate(value = case_when(item %in% head_1000 ~ value*1000,
                             TRUE ~ value)) %>% 
    #all units are now in head
    mutate(unit = "Head") %>% 
    #get rid of superfluous columns
    select(area, item, group, year, value, unit)
)


food_items <- c("Eggs, hen, in shell" , "Eggs, other bird, in shell", "Meat nes", "Meat, ass" ,"Meat, bird nes","Meat, buffalo", "Meat, camel", "Meat, cattle", "Meat, chicken", "Meat, duck","Meat, game","Meat, goat", "Meat, goose and guinea fowl", "Meat, horse","Meat, mule", "Meat, other camelids"       
, "Meat, other rodents", "Meat, pig", "Meat, rabbit", "Meat, sheep", "Meat, turkey", "Milk, whole fresh buffalo", "Milk, whole fresh camel", "Milk, whole fresh cow", "Milk, whole fresh goat", "Milk, whole fresh sheep", "Offals, edible, buffaloes", "Offals, edible, camels", "Offals, edible, cattle", "Offals, edible, goats", "Offals, horses", "Offals, pigs, edible", "Offals, sheep,edible")

equine_meat <- c("Meat, ass" , "Meat, horse",  "Meat, mule")
bird_meat <- c("Meat, turkey", "Meat, duck", "Meat, game", "Meat, goose and guinea fowl", "Meat, bird nes")
camelid_meat <- c("Meat, camel", "Meat, other camelids")
other_meat <- c("Meat, rabbit", "Meat, other rodents", "Meat nes")
eggs <- c("Eggs, hen, in shell" , "Eggs, other bird, in shell")
milk <- c("Milk, whole fresh buffalo", "Milk, whole fresh camel", "Milk, whole fresh cow", "Milk, whole fresh goat", "Milk, whole fresh sheep")
offals <- c("Offals, edible, buffaloes", "Offals, edible, camels", "Offals, edible, cattle", "Offals, edible, goats", "Offals, horses", "Offals, pigs, edible", "Offals, sheep,edible")
bovine_meat <- c("Meat, cattle", "Meat, buffalo")
ovine_meat <- c("Meat, goat", "Meat, sheep")



(
  production_tidy <- livestock_primary %>% 
    filter(element=="Production" & unit == "tonnes" & item %in% food_items) %>% 
    mutate(group = case_when(item %in% equine_meat ~ "Equine meat",
                             item %in% bird_meat  ~ "Poultry & game meat, other",
                             item %in% camelid_meat ~ "Camelid meat",
                             item %in% other_meat  ~ "Miscellaneous meats",
                             item %in% eggs ~ "Eggs",
                             item %in% milk ~ "Milk", 
                             item %in% offals ~ "Offals",
                             item %in% ovine_meat ~ "Sheep & goat meat",
                             item %in% bovine_meat ~ "Cattle and buffalo meat",
                             item == "Meat, pig" ~ "Pig meat",
                             item == "Meat, chicken" ~ "Chicken meat",
                             TRUE ~ item)) %>% 
    select(area, item, group, element, year, value, unit)
  
  
)


```


## Select the countries we want to retain

```{r}

#sort coding out for both stocks and production 

Encoding(stocks_tidy$area) <- "latin1" #deals with the non-UTF
stocks_tidy$area <- iconv(stocks_tidy$area, "latin1", "UTF-8",sub='')

Encoding(production_tidy$area) <- "latin1" #deals with the non-UTF
production_tidy$area <- iconv(production_tidy$area, "latin1", "UTF-8",sub='')


#Composite countries pre-dissolution to highlight and combine - main shocks detected here will be geopolitical however

ussr <- c("Armenia", "Azerbaijan", "Belarus", "Estonia", "Georgia", "Kazakhstan", "Kyrgyzstan", "Latvia" , "Lithuania", "Republic of Moldova", "Russian Federation" , "Tajikistan", "Turkmenistan" , "Ukraine"  , "USSR" , "Uzbekistan" )

yugoslavia <- c("Bosnia and Herzegovina", "Croatia", "Montenegro", "Serbia", "Serbia and Montenegro", "Slovenia", "North Macedonia", "Yugoslav SFR")

czechoslovakia <- c("Czechia", "Czechoslovakia", "Slovakia" )

sudan <- c("South Sudan", "Sudan", "Sudan (former)")

ethiopia <- c("Eritrea", "Ethiopia" , "Ethiopia PDR" )


stocks_tidy_2 <- 
  stocks_tidy %>% 
  mutate(area_2 = case_when(area %in% ussr ~ "Former USSR",
                            area %in% yugoslavia ~ "Former Yugoslavia",
                            area %in% czechoslovakia ~ "Former Czechoslovakia",
                            area %in% sudan ~ "Former Sudan",
                            area %in% ethiopia ~ "Former Ethiopia",
                            TRUE ~ area))


production_tidy_2 <- 
  production_tidy %>% 
  mutate(area_2 = case_when(area %in% ussr ~ "Former USSR",
                            area %in% yugoslavia ~ "Former Yugoslavia",
                            area %in% czechoslovakia ~ "Former Czechoslovakia",
                            area %in% sudan ~ "Former Sudan",
                            area %in% ethiopia ~ "Former Ethiopia",
                            TRUE ~ area))



#find the country and item combinations that do not have a full time_series

(
  stocks_tally <- stocks_tidy_2 %>% 
    filter(!is.na(value)) %>% 
    group_by(area, item) %>% 
    summarise(n = n())
)


(
  production_tally <- production_tidy_2 %>% 
    filter(!is.na(value)) %>% 
    group_by(area, item) %>% 
    summarise(n = n())
)



#join by the length of non-NA time series,  remove continents, and join broken states

(
  stocks_tidy_3 <- 
    stocks_tidy_2 %>% 
    left_join(stocks_tally, by= c("area", "item")) %>%
    mutate(iso_3c = countrycode(area, origin = "country.name", destination = "iso3c", warn=TRUE)) %>% 
    mutate(iso_3c = case_when(area %in% c("Serbia and Montenegro", "Yugoslav FR",  "Czechoslovakia") ~ "FMR",
                              TRUE ~ iso_3c)) %>% 
    filter(!is.na(iso_3c))
)
  
  
(
  production_tidy_3 <- 
    production_tidy_2 %>% 
    left_join(production_tally, by= c("area", "item"))%>% 
    mutate(iso_3c = countrycode(area, origin = "country.name", destination = "iso3c", warn=TRUE)) %>% 
    mutate(iso_3c = case_when(area %in% c("Serbia and Montenegro", "Yugoslav FR",  "Czechoslovakia") ~ "FMR",
                              TRUE ~ iso_3c)) %>% 
    filter(!is.na(iso_3c))
)


```


## Get the country groupings we want for the analysis
```{r, include=FALSE}

# former_countries_stocks <- stocks_tidy_3 %>% 
#   filter(area %in% c(ussr, yugoslavia, czechoslovakia, sudan, ethiopia)) %>% 
#   mutate(value = case_when(is.na(value) ~ 0,
#                            TRUE ~ value)) %>% 
#   ungroup() %>% 
#   group_by(area_2, group, year) %>% 
#   summarise(value)
# 
# former_countries_production <- production_tidy_3 %>% 
#   filter(area %in% c(ussr, yugoslavia, czechoslovakia, sudan, ethiopia)) %>% 
#   mutate(value = case_when(is.na(value) ~ 0,
#                            TRUE ~ value)) %>% 
#   group_by(iso_3c, area_2, item,year, unit) %>% 
#   summarise(value)



stocks_tidy_no_former <- 
  stocks_tidy_3 %>%
  filter(n == 59) %>% 
  group_by(iso_3c, area, group,year, unit) %>% 
  summarise(value = sum(value)) %>% 
  saveRDS(here("data", "production_tidy","stocks_wo_former_nations.rds"))


production_tidy_no_former <- 
  production_tidy_3 %>% 
  filter(n == 59) %>% 
  group_by(iso_3c, area, group, year, unit) %>% 
  summarise(value = sum(value)) %>% 
  saveRDS(here("data", "production_tidy","production_wo_former_nations.rds"))
  


#rm(production_tidy,production_tidy_2, production_tidy_3, stocks_tidy, stocks_tidy_2, stocks_tidy_3)

```
