==> scripts/2b_testing_shock_detection_function.Rmd <==
---
title: "Shock detection function"
output: html_document
---

## Libraries and setup
```{r setup, include=FALSE}

library(here) 
library(tidyverse)

```

##Function for shock detection
```{r}

source(here("scripts", "2_shock_detection_function.R"))


some_test_data <- tibble(readRDS(here("data", "production_tidy", "production_wo_former_nations.rds"))) %>% slice(1:59) 


shock_years <- some_test_data$year[which(shock_detector_gam(target_time_series = some_test_data$value, target_cooks_distance = 0.3) == 1)]#shock years

#looks good
ggplot(some_test_data, aes(x=year, y=value))+
  geom_line()+
  geom_vline(xintercept = shock_years, linetype="dashed", colour="red")



```
