==> scripts/4_stocks_shock_detection.Rmd <==
---
title: "Shock detection in animal stocks"
output: html_document
---

#libraries
```{r setup, include=FALSE}
library(here)
library(vroom)
library(tidyverse)
library(zoo)



source(here("scripts", "2_shock_detection_function.R"))
```

## Import tidied production data

```{r cars}
production <- tibble(readRDS(here("data", "production_tidy", "stocks_wo_former_nations.rds"))) %>% ungroup()


```

## Set up data for different produciton baselines (duration and mean or median) - no need to adjust span (GAM function far better at fitting first time than loess of set span).

```{r pressure, echo=FALSE}

#add baseline durations

baseline_durations <- c(3,5,7)

list_of_tibbles <- vector(mode="list", length = 3)

for(duration in baseline_durations){
  this_duration <- 
  production %>% 
  group_by(iso_3c, area, group) %>% 
  nest() %>% 
  mutate(baseline_durations = duration,
         mean = map(data, ~(rollapply(data = .$value, width = duration, FUN = mean, fill=NA, align = "right"))),
         median = map(data, ~(rollapply(data = .$value, width = duration, FUN = median, fill=NA, align = "right")))
  ) %>% 
  unnest()

  
list_of_tibbles[[duration]] <- this_duration
}



production_w_baselines <- 
  bind_rows(list_of_tibbles) %>% 
  pivot_longer(names_to = "average", values_to = "lagged_value",
               -c(iso_3c, area, group, year, unit, value, baseline_durations)) %>% 
  arrange(area, group, baseline_durations, average,  year)


```

## Detect shocks with GAM varying cooks distance
```{r}

#rules of thumb for cooks distance applicable to each time series
# 4/n = 4/59 = ~0.07
# Greater than 1 
#do a sequence between the two



#using Gam shock detector

cooks_distances <- seq(from =0.05, to=1, by=0.05)

shocks_gam_apply <- function(this_cd){
  
  message(paste("Processing...cooks' distance =", this_cd))
  
  #data frame with shock
  these_shocks <- 
    production_w_baselines %>% 
    group_by(iso_3c, area, group,unit, baseline_durations, average) %>% 
    nest() %>% 
    mutate(shock_presence = map(data, ~shock_detector_gam(target_time_series = .$value, target_cooks_distance = this_cd))) %>% 
    unnest(cols = c(data, shock_presence)) %>% 
    mutate(cooks_distance = this_cd)
  
  return(these_shocks)
  
  
}
  
shocks_by_cooks_distance <- map_df(cooks_distances, shocks_gam_apply) %>% 
  mutate(delta_production = lagged_value - value) 

#export shocks data

saveRDS(shocks_by_cooks_distance, file = here("data", "generated_data", "shocks_to_stocks", "shocks_w_baselines_cooks_distance_STOCKS.rds"))




```

## Sensitivity analysis around cooks's distance

```{r}

#0.3 - 0.35 seems reasonable for where a small blip in the trajectory starts

sensitivity <- shocks_by_cooks_distance %>%
  filter(shock_presence==1) %>% 
  drop_na(lagged_value) %>% 
  group_by(cooks_distance) %>% 
  summarise(n = n()) %>% 
  mutate(diff = c(0, diff(n))/lag(n))
  

ggplot(data = sensitivity, aes(x=cooks_distance, y=n))+geom_bar(stat = "identity")+geom_vline(lty=2, xintercept = 0.35)


```


## Detect shocks with LOESS varying span
```{r}


#range of reaosnable spans for LOESS models

spans <- seq(from =0.4, to=0.8, by=0.1)

shock_loess_apply <-   function(this_span){
  
  
  #data frame with shock
  these_shocks <- 
    production_w_baselines %>% 
    group_by(iso_3c, area, group, baseline_durations, average) %>% 
    nest() %>% 
    mutate(shock_presence = map(data, ~shock_detector_loess(target_time_series = .$value, target_cooks_distance = 0.3, target_span = this_span))) %>% 
    unnest(cols = c(data, shock_presence)) %>% 
    mutate(span = this_span)
  
  return(these_shocks)
}
  

shocks_by_span <- map_df(spans, shock_loess_apply) %>% 
  mutate(delta_production = lagged_value - value) %>% 
  filter(shock_presence==1)

#export shocks data

saveRDS(shocks_by_span, file = here("data", "generated_data",  "shocks_to_stocks", "shocks_w_baselines_span_STOCKS.rds"))


```



#span sensitivity
```{r}

sensitivity_2 <- shocks_by_span %>% 
  filter(shock_presence==1) %>% 
  drop_na(lagged_value) %>% 
  group_by(span) %>% 
  summarise(n = n())

ggplot(sensitivity_2)+
  aes(x=span, y=n)+
  geom_line()

```
